Add a circuit builder
This circuit builder requires exact number of inputs of gates, but I guess that is good in a way of clarity. Also, it defaults to Qiskit endianness. But with the `assembler` parameter we can turn it to follow native endian quite easily.
An example gate for 2 bit encoding:
```
(define c1 (make-circuit (list (list hadamard
                                       (I 2))
                                 qCX
                                 (list pauli-z
                                       (I 2))
                                 (list pauli-x
                                       (I 2))
                                 qCX
                                 (list hadamard
                                       (I 2)))))
```